chore: move menubutton to sep. component
