+++ src/components/Settings/index.tsx
@@ -1,53 +1,17 @@
 // eslint-disable-next-line no-restricted-imports
-import { t } from '@lingui/macro'
 import { Percent } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
 import { AutoColumn } from 'components/Column'
-import Row from 'components/Row'
 import { L2_CHAIN_IDS } from 'constants/chains'
 import { isSupportedChainId } from 'lib/hooks/routing/clientSideSmartOrderRouter'
-import { useRef } from 'react'
-import { Settings } from 'react-feather'
-import { useModalIsOpen, useToggleSettingsMenu } from 'state/application/hooks'
+import { useModalIsOpen } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
-import { useUserSlippageTolerance } from 'state/user/hooks'
-import { SlippageTolerance } from 'state/user/types'
 import styled from 'styled-components/macro'
-import { ThemedText } from 'theme'
-import validateUserSlippageTolerance, { SlippageValidationResponse } from 'utils/validateUserSlippageTolerance'
 
 import MaxSlippageSettings from './MaxSlippageSettings'
+import MenuButton from './MenuButton'
 import RouterPreferenceSettings from './RouterPreferenceSettings'
 import TransactionDeadlineSettings from './TransactionDeadlineSettings'
-
-const MenuIcon = styled(Settings)`
-  height: 20px;
-  width: 20px;
-  > * {
-    stroke: ${({ theme }) => theme.textSecondary};
-  }
-`
-
-const MenuButton = styled.button<{ disabled: boolean }>`
-  position: relative;
-  width: 100%;
-  height: 100%;
-  border: none;
-  background-color: transparent;
-  margin: 0;
-  padding: 0;
-  border-radius: 0.5rem;
-  ${({ disabled }) =>
-    !disabled &&
-    `
-    :hover,
-    :focus {
-      cursor: pointer;
-      outline: none;
-      opacity: 0.7;
-    }
-  `}
-`
 
 const Menu = styled.div`
   margin-left: 0.5rem;
@@ -89,61 +53,15 @@ const Divider = styled.div`
   background-color: ${({ theme }) => theme.backgroundOutline};
 `
 
-const MenuIconContainer = styled(Row)`
-  padding: 6px 12px;
-  border-radius: 16px;
-`
-
-const MenuIconContainerWithSlippage = styled(MenuIconContainer)<{ displayWarning?: boolean }>`
-  div {
-    color: ${({ theme, displayWarning }) => (displayWarning ? theme.accentWarning : theme.textSecondary)};
-  }
-
-  background-color: ${({ theme, displayWarning }) =>
-    displayWarning ? theme.accentWarningSoft : theme.backgroundModule};
-`
-
-const MenuButtonIcon = () => {
-  const [userSlippageTolerance] = useUserSlippageTolerance()
-
-  if (userSlippageTolerance === SlippageTolerance.Auto) {
-    return (
-      <MenuIconContainer>
-        <MenuIcon />
-      </MenuIconContainer>
-    )
-  }
-
-  const isInvalidSlippage = validateUserSlippageTolerance(userSlippageTolerance) !== SlippageValidationResponse.Valid
-
-  return (
-    <MenuIconContainerWithSlippage gap="sm" displayWarning={isInvalidSlippage}>
-      <ThemedText.Caption>{userSlippageTolerance.toFixed(2)}% slippage</ThemedText.Caption>
-      <MenuIcon />
-    </MenuIconContainerWithSlippage>
-  )
-}
-
 export default function SettingsTab({ autoSlippage }: { autoSlippage: Percent }) {
   const { chainId } = useWeb3React()
   const showDeadlineSettings = Boolean(chainId && !L2_CHAIN_IDS.includes(chainId))
 
-  const node = useRef<HTMLDivElement | null>(null)
   const open = useModalIsOpen(ApplicationModal.SETTINGS)
 
-  const toggle = useToggleSettingsMenu()
-
   return (
-    <Menu ref={node}>
-      <MenuButton
-        disabled={!isSupportedChainId(chainId)}
-        onClick={toggle}
-        id="open-settings-dialog-button"
-        data-testid="open-settings-dialog-button"
-        aria-label={t`Transaction Settings`}
-      >
-        <MenuButtonIcon />
-      </MenuButton>
+    <Menu>
+      <MenuButton />
       {open && (
         <MenuFlyout>
           <AutoColumn gap="16px" style={{ padding: '1rem' }}>
